refactor: move `IconButton` to the global components folder
